refactor: splits off virtual code owners anomaly processing code
