+++ src/read-virtual-code-owners.ts
@@ -1,10 +1,8 @@
 import { readFileSync } from "node:fs";
 import { EOL } from "node:os";
 import type { IAnomaly, ITeamMap, IVirtualCodeOwnersCST } from "types/types.js";
-import {
-  getAnomalies,
-  parse as parseVirtualCodeOwners,
-} from "./parse-virtual-code-owners.js";
+import { parse as parseVirtualCodeOwners } from "./parse-virtual-code-owners.js";
+import { getAnomalies } from "./virtual-code-owners-anomalies.js";
 
 export default function readVirtualCodeOwners(
   pVirtualCodeOwnersFileName: string,
